Splitting up depparser code See #69
