+++ src/ipbb/depparser/_depparser_core.py
@@ -12,87 +12,9 @@ from collections import OrderedDict
 from os.path import exists, splitext
 from string import Template
 
+from ._cmdparser import ComponentAction, DepCmdParser, DepCmdParserError
+from ._cmdtypes import FileCommand, IncludeCommand
 
-# -----------------------------------------------------------------------------
-class Command(object):
-    """Container class for dep commands parsed form dep files
-
-    Attributes:
-        cmd       (str): command directive
-        FilePath  (str): absolute, normalised path to the command target.
-        Package   (str): package the target belongs to.
-        Component (str): component withon 'Package' the target belongs to
-    """
-
-    def __init__(self, aCmd, aFilePath, aPackage, aComponent):
-        super(Command, self).__init__()
-        self.cmd = aCmd
-        self.FilePath = aFilePath
-        self.Package = aPackage
-        self.Component = aComponent
-
-    def __str__(self):
-        return '{ \'{}\', component: \'{}:{}\' }' % (
-            self.cmd, self.FilePath, self.Package, self.Component
-        )
-
-
-# -----------------------------------------------------------------------------
-class FileCommand(Command):
-    """Container class for dep commands parsed form dep files
-
-    Attributes:
-        cmd       (str):  command directive
-        FilePath  (str):  absolute, normalised path to the command target.
-        Package   (str):  package the target belongs to.
-        Component (str):  component withon 'Package' the target belongs to
-        Lib       (str):  library the file will be added to
-        TopLevel  (bool): addrtab-only flag, identifies address table as top-level
-        Vhdl2008  (bool): src-only flag, toggles the vhdl 2008 syntax for .vhd files
-        Finalise  (bool): setup-only flag, identifies setup scripts to be executed at the end
-
-    """
-    # --------------------------------------------------------------
-    def __init__(self, aCmd, aFilePath, aPackage, aComponent, aLib, aTopLevel, aVhdl2008, aFinalise):
-        super(FileCommand, self).__init__(aCmd, aFilePath, aPackage, aComponent)
-
-        self.Lib = aLib
-        self.TopLevel = aTopLevel
-        self.Vhdl2008 = aVhdl2008
-        self.Finalise = aFinalise
-
-    # --------------------------------------------------------------
-    def __str__(self):
-
-        lFlags = self.flags()
-        return '{ \'%s\', flags: %s, component: \'%s:%s\' }' % (
-            self.FilePath, ''.join(lFlags) if lFlags else 'none', self.Package, self.Component
-        )
-
-    # --------------------------------------------------------------
-    def flags(self):
-        lFlags = []
-        if self.TopLevel:
-            lFlags.append('top')
-        if self.Vhdl2008:
-            lFlags.append('vhdl2008')
-        if self.Finalise:
-            lFlags.append('finalise')
-        return lFlags
-
-    __repr__ = __str__
-
-    # --------------------------------------------------------------
-    def __eq__(self, other):
-        return (self.FilePath == other.FilePath) and (self.Lib == other.Lib)
-
-
-# -----------------------------------------------------------------------------
-class IncludeCommand(Command):
-    """docstring for IncludeCommand"""
-    def __init__(self, aCmd, aFilePath, aPackage, aComponent, aDepFileObj=None):
-        super(IncludeCommand, self).__init__(aCmd, aFilePath, aPackage, aComponent)
-        self.depfile = aDepFileObj
 
 
 # -----------------------------------------------------------------------------
@@ -137,111 +59,6 @@ class DepFile(object):
         for f in self.children:
             for sf in f.iterchildren():
                 yield sf
-
-
-# -----------------------------------------------------------------------------
-class ComponentAction(argparse.Action):
-    '''
-    Parses <module>:<component>
-    '''
-
-    def __call__(self, parser, namespace, values, option_string=None):
-        lSeparators = values.count(':')
-        # Validate the format
-        if lSeparators > 1:
-            raise argparse.ArgumentTypeError(
-                'Malformed component name : %s. Expected <module>:<component>' % values)
-
-        lTokenized = values.split(':')
-        if len(lTokenized) == 1:
-            lTokenized.insert(0, None)
-
-        setattr(namespace, self.dest, tuple(lTokenized))
-
-
-# -----------------------------------------------------------------------------
-class DepCmdParserError(Exception):
-    pass
-
-# -----------------------------------------------------------------------------
-class SrcTypeAction(argparse.Action):
-    def __init__(self, *args, **kwargs):
-        super(SrcTypeAction, self).__init__(*args, **kwargs)
-        self._choices = ['synth', 'sim']
-        self.default = self._choices
-
-    def __call__(self, parser, namespace, values, option_string=None):
-
-        tokens = values.split(',')
-        
-        lInvalid = [t for t in tokens if t not in self._choices]
-        if lInvalid:
-            raise ValueError('Invalid source types '+','.join(lInvalid))
-
-        setattr(namespace, self.dest, tokens )
-
-
-# -----------------------------------------------------------------------------
-class DepCmdParser(argparse.ArgumentParser):
-    def error(self, message):
-        raise DepCmdParserError(message)
-
-    # ---------------------------------
-    def __init__(self, *args, **kwargs):
-        super(DepCmdParser, self).__init__(*args, **kwargs)
-
-        # Common options
-        lCompArgOpts = dict(action=ComponentAction, default=(None, None))
-
-        parser_add = self.add_subparsers(dest='cmd', parser_class=argparse.ArgumentParser)
-
-        # Include sub-parser
-        subp = parser_add.add_parser('include')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('--cd')
-        subp.add_argument('file', nargs='*')
-
-        # Setup sub-parser
-        subp = parser_add.add_parser('setup')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('--cd')
-        subp.add_argument('file', nargs='*')
-        subp.add_argument('-f', '--finalise', action='store_true')
-
-        # Utilites sub-parser
-        subp = parser_add.add_parser('util')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('--cd')
-        subp.add_argument('file', nargs='*')
-
-        # Source sub-parser
-        subp = parser_add.add_parser('src')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('-l', '--lib')
-        subp.add_argument('--cd')
-        subp.add_argument('file', nargs='+')
-        subp.add_argument('--vhdl2008', action='store_true')
-        subp.add_argument('-t', '--tyoe', action=SrcTypeAction)
-
-        # Address table sub-parser
-        subp = parser_add.add_parser('addrtab')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('--cd')
-        subp.add_argument('-t', '--toplevel', action='store_true')
-        subp.add_argument('file', nargs='*')
-
-        # Ip repository sub-parser
-        subp = parser_add.add_parser('iprepo')
-        subp.add_argument('-c', '--component', **lCompArgOpts)
-        subp.add_argument('--cd')
-        subp.add_argument('file', nargs='*')
-
-        # --------------------------------------------------------------
-
-    def parseLine(self, *args, **kwargs):
-
-        return self.parse_args(*args, **kwargs)
-# -----------------------------------------------------------------------------
 
 
 # -----------------------------------------------------------------------------
